backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
websockets==14.1

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Ollama integration
ollama==0.4.4

# Pinecone vector database
pinecone-client==5.0.1

# Environment and utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# For file processing
python-magic==0.4.27
pypdf==5.1.0
python-docx==1.1.2
